+++ lib/data/models/money_objects/splits/popup_transaction_splits.dart
@@ -1,11 +1,12 @@
+import 'package:get/get.dart';
 import 'package:money/core/widgets/gaps.dart';
 import 'package:money/core/widgets/widgets.dart';
 import 'package:money/data/models/money_objects/transactions/transaction.dart';
 import 'package:money/views/home/sub_views/adaptive_view/adaptive_list/transactions/list_view_transaction_splits.dart';
 
-void showTransactionSplits(final BuildContext context, final Transaction transaction) {
+void showTransactionSplits(final Transaction transaction) {
   adaptiveScreenSizeDialog(
-    context: context,
+    context: Get.context!,
     title: 'Transaction split',
     child: IntrinsicHeight(
       child: Column(